chore: dead noqa and small correctness fixes surfaced by qualety (#117)
Ran the qualety code-quality tool over the repo and applied the findings
that hold up against our own ruff config:

- 157 dead `# noqa: E402` directives removed. ruff exempts imports that
  follow a bare `sys.path` mutation, so the suppressions in the dev-script
  preambles never did anything. The 15 in `lab98_expert_recalibration_sweep.py`
  are kept — an assignment and a mid-block `def` break that exemption there,
  so E402 genuinely fires.
- `__enter__` on `_StatusLine` and `StereoHandSource` returned the concrete
  class, so `with` on a subclass inferred the base. Return `Self` (PYI034).
- Unused unpacked variables underscore-prefixed in four dev scripts (RUF059).
- `dict(...)` → dict literals (C408), `str(x)` in an f-string → `!s` (RUF010),
  redundant `int(round(...))` cast dropped (RUF046), `return None` → `return`
  in a None-returning branch (RET501), `import x as x` → `from y import x`
  (PLR0402), and two `for k in d: d[k]` loops → `.items()` (PLC0206).

ruff, mypy and the 296-test suite are green; the five touched dev scripts
still import cleanly.

+++ scripts/dev_harness_controller.py
@@ -31,14 +31,14 @@ import numpy as np
 # Allow running before the package is installed in the venv.
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent / "src"))
 
-from ai_teleop.common.command import Command  # noqa: E402
-from ai_teleop.common.log import (  # noqa: E402
+from ai_teleop.common.command import Command
+from ai_teleop.common.log import (
     add_logging_arguments,
     configure_from_args,
     get_logger,
 )
-from ai_teleop.control import Controller, LockState  # noqa: E402
-from ai_teleop.sim.scene import SimEnv  # noqa: E402
+from ai_teleop.control import Controller, LockState
+from ai_teleop.sim.scene import SimEnv
 
 REPO_ROOT = Path(__file__).resolve().parent.parent
 SCENE_PATH = REPO_ROOT / "assets" / "mjcf" / "full_scene.xml"
@@ -103,7 +103,7 @@ def run_phase(
     `target_pos_fn(t)` returns the commanded target position at sim time `t`
     (so phases can ramp). `target_quat` is held constant within a phase.
     """
-    n_steps = int(round(duration_s / SIM_DT))
+    n_steps = round(duration_s / SIM_DT)
     peak_force = 0.0
     last_pos_err = float("nan")
     lock_state_changes: list[tuple[float, str, str]] = []
